+++ packages/browser-wallet/src/shared/utils/verifiable-credential-helpers.ts
@@ -20,6 +20,8 @@ import { applyExecutionNRGBuffer, getContractName } from './contract-helpers';
 import { getNet } from './network-helpers';
 import { logError } from './log-helpers';
 
+export type VerifiableCredentialSchemaWithFallback = VerifiableCredentialSchema & { usingFallback: boolean };
+
 /**
  * Extracts the credential holder id from a verifiable credential id (did).
  * @param credentialId the did for a credential
@@ -49,7 +51,7 @@ export function getPublicKeyfromPublicKeyIdentifierDID(did: string) {
 }
 
 /**
- * Extracts the credential registry contract addres from a verifiable credential id (did).
+ * Extracts the credential registry contract address from a verifiable credential id (did).
  * @param credentialId the did for a credential
  * @returns the contract address of the issuing contract of the provided credential id
  */